autofeature-ai/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
aiohttp>=3.9.0
openai>=1.52.0
databricks-sdk>=0.30.0
databricks-sql-connector>=3.0.0
pydantic>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
featuretools>=1.28.0
woodwork>=0.28.0
